+++ AppSyncRealTimeClient/ConnectionProvider/AppsyncRealtimeConnection/RealtimeConnectionProvider+ConnectionInterceptable.swift
@@ -17,11 +17,6 @@ extension RealtimeConnectionProvider: ConnectionInterceptable {
         _ request: AppSyncConnectionRequest,
         for endpoint: URL
     ) -> AppSyncConnectionRequest {
-        guard let connectionInterceptors = connectionInterceptors as? [ConnectionInterceptor] else {
-            AppSyncLogger.error("Failed to cast messageInterceptors.")
-            return request
-        }
-
         let finalRequest = connectionInterceptors.reduce(request) { $1.interceptConnection($0, for: endpoint) }
         return finalRequest
     }
