+++ AppSyncRealTimeClient/ConnectionProvider/AppsyncRealtimeConnection/RealtimeConnectionProvider+MessageInterceptable.swift
@@ -14,11 +14,6 @@ extension RealtimeConnectionProvider: MessageInterceptable {
     }
 
     public func interceptMessage(_ message: AppSyncMessage, for endpoint: URL) -> AppSyncMessage {
-        guard let messageInterceptors = messageInterceptors as? [MessageInterceptor] else {
-            AppSyncLogger.error("Failed to cast messageInterceptors.")
-            return message
-        }
-
         let finalMessage = messageInterceptors.reduce(message) { $1.interceptMessage($0, for: endpoint) }
         return finalMessage
     }
