+++ AppSyncRealTimeClient/ConnectionProvider/AppsyncRealtimeConnection/RealtimeConnectionProviderAsync+ConnectionInterceptableAsync.swift
@@ -11,16 +11,10 @@ import Foundation
 extension RealtimeConnectionProviderAsync: ConnectionInterceptableAsync {
 
     public func addInterceptor(_ interceptor: ConnectionInterceptorAsync) {
-        useAsyncInterceptors = true
         connectionInterceptors.append(interceptor)
     }
 
     public func interceptConnection(_ request: AppSyncConnectionRequest, for endpoint: URL) async -> AppSyncConnectionRequest {
-        guard let connectionInterceptors = connectionInterceptors as? [ConnectionInterceptorAsync] else {
-            AppSyncLogger.error("Failed to cast messageInterceptors.")
-            return request
-        }
-
         var finalRequest = request
         for interceptor in connectionInterceptors {
             finalRequest = await interceptor.interceptConnection(finalRequest, for: endpoint)
