+++ AppSyncRealTimeClient/ConnectionProvider/AppsyncRealtimeConnection/RealtimeConnectionProviderAsync+MessageInterceptableAsync.swift
@@ -8,18 +8,12 @@
 import Foundation
 
 @available(iOS 13.0.0, *)
-extension RealtimeConnectionProviderBase: MessageInterceptableAsync {
+extension RealtimeConnectionProviderAsync: MessageInterceptableAsync {
     public func addInterceptor(_ interceptor: MessageInterceptorAsync) {
-        useAsyncInterceptors = true
         messageInterceptors.append(interceptor)
     }
 
     public func interceptMessage(_ message: AppSyncMessage, for endpoint: URL) async -> AppSyncMessage {
-        guard let messageInterceptors = messageInterceptors as? [MessageInterceptorAsync] else {
-            AppSyncLogger.error("Failed to cast messageInterceptors.")
-            return message
-        }
-
         var finalMessage = message
         for interceptor in messageInterceptors {
             finalMessage = await interceptor.interceptMessage(finalMessage, for: endpoint)
