--- AppSyncRealTimeClient/ConnectionProvider/AppsyncRealtimeConnection/RealtimeConnectionProviderAsync.swift
@@ -1,13 +1,0 @@
-//
-//  RealtimeConnectionProviderAsync.swift
-//  AppSyncRealTimeClient
-//
-//  Created by Ameter, Chris on 4/28/22.
-//  Copyright © 2022 amazonaws. All rights reserved.
-//
-
-import Foundation
-
-public class RealtimeConnectionProviderAsync: RealtimeConnectionProviderBase {
-    
-}
